Reorganizacion y capas

+++ src/controllers/product.Controller.js
@@ -1,8 +1,163 @@
-const Products = require('../dao/models/products.model');
+// const Products = require('../dao/models/products.model');
+
+// class ProductController {
+
+//     constructor(ProductService){
+//         this.service = ProductService;
+//     }
+
+//     #handleError(res, err) {
+//         if (err.message === 'not found') {
+//             return res.status(404).json({ error: 'Not found' });
+//         }
+
+//         if (err.message === 'invalid parameters') {
+//             return res.status(400).json({ error: 'Invalid parameters' });
+//         }
+
+//         return res.status(500).json({ error: err.message });
+//     }
+
+//     //Obtener listado de todos los productos o limitarlo por cantidad
+
+//     async getAll(req, res){
+//         try {
+//             const isLoggedIn = ![null, undefined].includes(req.session.user);
+//             const user = req.session.user;
+
+//             const cartId = req.session.cartId;
+
+//             const limit = parseInt(req.query.limit) || 10;
+//             const page = parseInt(req.query.page) || 1;
+//             const sort = req.query.sort === 'desc' ? -1 : 1;
+//             const query = req.query.query || '';
+//             const category = req.query.category || '';
+//             const availability = req.query.availability || '';
+    
+//             const match = {};
+//             if (query) {
+//                 match.$or = [
+//                     { category: { $regex: query, $options: 'i' } },
+//                     { availability: { $regex: query, $options: 'i' } }
+//                 ];
+//             }
+//             if (category) {
+//                 match.category = { $regex: category, $options: 'i' };
+//             }
+//             if (availability) {
+//                 match.availability = { $regex: availability, $options: 'i' };
+//             }
+    
+//             const totalProducts = await Products.countDocuments(match).lean();
+//             const totalPages = Math.ceil(totalProducts / limit);
+    
+//             const skip = (page - 1) * limit;
+    
+//             const products = await Products.find(match).lean()
+//                 .sort({ price: sort })
+//                 .skip(skip)
+//                 .limit(limit);
+    
+//             const hasNextPage = page < totalPages;
+//             const hasPrevPage = page > 1;
+    
+//             const prevPage = hasPrevPage ? page - 1 : null;
+//             const nextPage = hasNextPage ? page + 1 : null;
+    
+//             const prevLink = hasPrevPage ? `/products?page=${prevPage}&limit=${limit}&sort=${req.query.sort}&query=${query}&category=${category}&availability=${availability}` : null;
+//             const nextLink = hasNextPage ? `/products?page=${nextPage}&limit=${limit}&sort=${req.query.sort}&query=${query}&category=${category}&availability=${availability}` : null;
+    
+//             res.render('home', {
+//                 title: 'Lista Productos',
+//                 products,
+//                 user: user,
+//                 cartId: user.cartId,
+//                 styles: ['style'],
+//                 useWS: false,
+//                 scripts: ['index'],
+//                 totalPages,
+//                 prevPage,
+//                 nextPage,
+//                 page,
+//                 hasPrevPage,
+//                 hasNextPage,
+//                 prevLink,
+//                 nextLink,
+//                 isLoggedIn,
+//                 isNotLoggedIn: !isLoggedIn
+//             });
+//         } catch (error) {
+//             return this.#handleError(res, error);
+//         }
+//     }
+
+//     //Obtener productos por ID
+
+//     async getById(req, res) {
+//         const productId = parseInt(req.params.id);
+
+//         try {
+//             const product = await this.service.getById({ id: productId });
+//             if (!product) {
+//                 res.status(404).json({ error: 'Producto no encontrado' });
+//             } else {
+//                 res.json(product);
+//             }
+//         } catch (error) {
+//             return this.#handleError(res, error);
+//         }
+//     }
+
+//     //Borrar producto por ID
+
+//     async deleteById(req, res){
+//         const productId = req.params.id;
+//         try {
+//             await this.service.deleteById(productId);
+//             res.json({ message: `Producto con ID ${productId} eliminado correctamente` });
+//         } catch (error) {
+//             return this.#handleError(res, error);
+//         }
+//     }
+
+//     //Cargar productos
+
+//     async createOne(req, res){
+//         const { title, description, price, thumbnail, code, stock, category } = req.body;
+        
+//         try {
+//             const producto = await this.service.createOne(title, description, price, thumbnail, code, stock, category);
+//             return res.json(producto);
+//         } catch (error) {
+//             return this.#handleError(res, error);
+//         }
+//     }
+
+//     //Actualizar producto por ID
+
+//     async updateOne(req, res){
+//         const productId = req.params.id;
+//         const updatedFields = req.body;
+//         try {
+//             const updatedProduct = await this.service.updateById(productId, updatedFields, { new: true });
+    
+//             if (!updatedProduct) {
+//                 return res.status(404).json({ error: 'Producto no encontrado' });
+//             }
+    
+//             res.json(updatedProduct);
+//         } catch (error) {
+//             return this.#handleError(res, error);
+//         }
+//     }
+// }
+
+// module.exports = { ProductController };
+
+const mongoose = require('mongoose');
 
 class ProductController {
-
-    constructor(ProductService){
+    constructor(ProductService) {
         this.service = ProductService;
     }
 
@@ -18,9 +173,7 @@ class ProductController {
         return res.status(500).json({ error: err.message });
     }
 
-    //Obtener listado de todos los productos o limitarlo por cantidad
-
-    async getAll(req, res){
+    async getAll(req, res) {
         try {
             const isLoggedIn = ![null, undefined].includes(req.session.user);
             const user = req.session.user;
@@ -33,7 +186,7 @@ class ProductController {
             const query = req.query.query || '';
             const category = req.query.category || '';
             const availability = req.query.availability || '';
-    
+
             const match = {};
             if (query) {
                 match.$or = [
@@ -47,26 +200,23 @@ class ProductController {
             if (availability) {
                 match.availability = { $regex: availability, $options: 'i' };
             }
-    
-            const totalProducts = await Products.countDocuments(match).lean();
+
+            const totalProducts = await this.service.countDocuments(match);
             const totalPages = Math.ceil(totalProducts / limit);
-    
+
             const skip = (page - 1) * limit;
-    
-            const products = await Products.find(match).lean()
-                .sort({ price: sort })
-                .skip(skip)
-                .limit(limit);
-    
+
+            const products = await this.service.find(match, { sort, skip, limit });
+
             const hasNextPage = page < totalPages;
             const hasPrevPage = page > 1;
-    
+
             const prevPage = hasPrevPage ? page - 1 : null;
             const nextPage = hasNextPage ? page + 1 : null;
-    
+
             const prevLink = hasPrevPage ? `/products?page=${prevPage}&limit=${limit}&sort=${req.query.sort}&query=${query}&category=${category}&availability=${availability}` : null;
             const nextLink = hasNextPage ? `/products?page=${nextPage}&limit=${limit}&sort=${req.query.sort}&query=${query}&category=${category}&availability=${availability}` : null;
-    
+
             res.render('home', {
                 title: 'Lista Productos',
                 products,
@@ -91,13 +241,15 @@ class ProductController {
         }
     }
 
-    //Obtener productos por ID
-
     async getById(req, res) {
-        const productId = parseInt(req.params.id);
+        const productId = req.params.id;
+
+        if (!mongoose.Types.ObjectId.isValid(productId)) {
+            return res.status(400).json({ error: 'Invalid product ID' });
+        }
 
         try {
-            const product = await this.service.getById({ id: productId });
+            const product = await this.service.getById(productId);
             if (!product) {
                 res.status(404).json({ error: 'Producto no encontrado' });
             } else {
@@ -108,10 +260,13 @@ class ProductController {
         }
     }
 
-    //Borrar producto por ID
-
-    async deleteById(req, res){
+    async deleteById(req, res) {
         const productId = req.params.id;
+
+        if (!mongoose.Types.ObjectId.isValid(productId)) {
+            return res.status(400).json({ error: 'Invalid product ID' });
+        }
+
         try {
             await this.service.deleteById(productId);
             res.json({ message: `Producto con ID ${productId} eliminado correctamente` });
@@ -120,11 +275,9 @@ class ProductController {
         }
     }
 
-    //Cargar productos
-
-    async createOne(req, res){
+    async createOne(req, res) {
         const { title, description, price, thumbnail, code, stock, category } = req.body;
-        
+
         try {
             const producto = await this.service.createOne(title, description, price, thumbnail, code, stock, category);
             return res.json(producto);
@@ -133,18 +286,21 @@ class ProductController {
         }
     }
 
-    //Actualizar producto por ID
-
-    async updateOne(req, res){
+    async updateOne(req, res) {
         const productId = req.params.id;
         const updatedFields = req.body;
+
+        if (!mongoose.Types.ObjectId.isValid(productId)) {
+            return res.status(400).json({ error: 'Invalid product ID' });
+        }
+
         try {
-            const updatedProduct = await this.service.updateById(productId, updatedFields, { new: true });
-    
+            const updatedProduct = await this.service.updateById(productId, updatedFields);
+
             if (!updatedProduct) {
                 return res.status(404).json({ error: 'Producto no encontrado' });
             }
-    
+
             res.json(updatedProduct);
         } catch (error) {
             return this.#handleError(res, error);
